매뉴얼 일부 작성 완료, 새 태그정보 init 파일에 분리

+++ src/pages/setting_page/setting_user/WithdrawConfirmPopup.tsx
@@ -1,9 +1,11 @@
-import React from "react";
+import React, { useRef } from "react";
 import { User } from "firebase/auth";
 
 import Popup from "../../../components/Popup";
 import Text from "../../../components/Text";
 import { FbAuth } from "../../../firebase/firebase_auth_service";
+import InputText from "../../../components/InputText";
+import { ColBox } from "../../../components/FlexBox";
 
 
 interface IWithdrawConfirmPopup {
@@ -14,12 +16,19 @@ interface IWithdrawConfirmPopup {
 
 const WithdrawConfirmPopup = ( { fbAuth, user, setIsOpenWithdrawPopup} : IWithdrawConfirmPopup ) => {
 
+  const confrimTextRef = useRef<HTMLTextAreaElement>(null)
+
   // 탈퇴 실행
   const doWithdrawUser = async () => {
-    await fbAuth.logOut();
+    if (confrimTextRef.current!.value !== "확인하였습니다") {
+      alert("올바른 확인 메세지를 입력해주세요")
+      setIsOpenWithdrawPopup(false)
+      return
+    }
     await fbAuth.withdrawUser();
     await fbAuth.deleteAllMemo();
     await fbAuth.deleteAllTag();
+    await fbAuth.logOut();
     
     alert("탈퇴 및 데이터 삭제가 완료되었습니다.")
   }
@@ -32,6 +41,7 @@ const WithdrawConfirmPopup = ( { fbAuth, user, setIsOpenWithdrawPopup} : IWithdr
 
   return(
     <Popup
+      gap={.1}
       title="회원 탈퇴"
       onClickCancel={ cancelWithdraw}
       onClickDo={() => doWithdrawUser()}
@@ -41,6 +51,22 @@ const WithdrawConfirmPopup = ( { fbAuth, user, setIsOpenWithdrawPopup} : IWithdr
         데이터 복구가 불가능합니다. <br />
         정말로 탈퇴할까요? <br />
       </Text>
+
+      <ColBox padding=".5rem" gap={.25}>
+        <Text padding="0 .25rem"> 
+          탈퇴시 아래 빈 칸에
+        </Text>
+        <Text bold padding="0 .25rem">
+          확인하였습니다
+        </Text>
+        <Text padding="0 .25rem">
+          를 입력해주세요.
+        </Text>
+        <InputText
+          ref={confrimTextRef}
+          placeholder="확인하였습니다" 
+        />
+      </ColBox>
     </Popup>
   )
 }
